model: Add irrigation, financial init, soil and crop types,
This commit introduces several major changes to the 'agents', 'model' and 'objects' Python files to enhance the capabilities of the model.

In 'agents.py', the Farmer class constructor (__init__ method) is significantly altered. The Farmer now has a money, yearly_living_expenses, lands, and neighbors attributes. The money and yearly_living_expenses attributes have been introduced to simulate the financial situation of the farmer, while the lands attribute, which is a list of FarmLand objects, is included to represent the farmer's lands. The neighbors attribute represents the farmer's social network.

In 'model.py', the AgriModel class constructor is refactored to accommodate the changes in the Farmer class and to further enhance the capabilities of the model. The model now includes the use of StagedActivation (instead of RandomActivation), which divides each step into several stages, thereby allowing to model different seasons in Indian agriculture. The soil_types_distribution dictionary and the crop_types list were added to better simulate the various soil and crop types present in the region. Each farmer now has multiple lands, and the land is created using a helper function, create_farmland().

In 'objects.py', the FarmLand class is also significantly modified. It now contains the attributes owner, size, monetary_value, irrigation_systems and current_crop. The Crop class is adjusted to match the changes in the Farmer and FarmLand classes. It now includes attributes sowing_density, allowed_irigation_types and the farm_land_planted. A new class, IrrigationSystem, is added to model different types of irrigation systems.

Note: Data collection is temporarily commented out in the model and will be re-enabled in a future commit.